feature(green): Rework of permissions to implement a RBAC + ABAC + split of permission file in multiple files (permission, rbac, predicate)
